refactor(agents): remove obsolete standalone agent files and consolidate into pipeline
