--- backend/internal/agents/enricher_test.go
@@ -1,136 +1,0 @@
-package agents
-
-import (
-	"database/sql"
-	"encoding/json"
-	"fmt"
-	"net/http"
-	"net/http/httptest"
-	"os"
-	"path/filepath"
-	"strings"
-	"testing"
-
-	"example.com/backend/internal/repository"
-	"go.uber.org/zap"
-	"gopkg.in/yaml.v3"
-	"gorm.io/driver/sqlite"
-	"gorm.io/gorm"
-	_ "modernc.org/sqlite"
-)
-
-func TestProcessEnrichFrontmatter_Success(t *testing.T) {
-	tempDir := t.TempDir()
-	articlesDir := filepath.Join(tempDir, "articles")
-	os.MkdirAll(articlesDir, 0755)
-
-	sqlDB, err := sql.Open("sqlite", filepath.Join(tempDir, "test.db"))
-	if err != nil {
-		t.Fatal(err)
-	}
-	db, err := gorm.Open(sqlite.Dialector{Conn: sqlDB}, &gorm.Config{})
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	db.AutoMigrate(&repository.GormArticle{})
-	db.Create(&repository.GormArticle{ID: 301, Title: "Go Concurrency Patterns"})
-
-	articlePath := filepath.Join(articlesDir, "301.md")
-	initialContent := "---\ntitle: Old Title\n---\n\nGo channels and goroutines make concurrent programming clean and safe."
-	os.WriteFile(articlePath, []byte(initialContent), 0644)
-
-	var capturedResponseFormat map[string]interface{}
-	ts := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		var req map[string]interface{}
-		json.NewDecoder(r.Body).Decode(&req)
-		if rf, ok := req["response_format"].(map[string]interface{}); ok {
-			capturedResponseFormat = rf
-		}
-
-		res := map[string]interface{}{
-			"choices": []map[string]interface{}{
-				{
-					"message": map[string]string{
-						"content": `{
-							"type": "Playbook",
-							"title": "Mastering Go Concurrency",
-							"description": "A comprehensive playbook for Go channels and goroutines.",
-							"resource": "https://golang.org/doc",
-							"tags": ["Golang", "Concurrency", "Channels"]
-						}`,
-					},
-				},
-			},
-		}
-		json.NewEncoder(w).Encode(res)
-	}))
-	defer ts.Close()
-
-	pool := &AgentPool{
-		Queue:         make(chan Job, 10),
-		logger:        zap.NewNop(),
-		db:            db,
-		dataDirectory: tempDir,
-	}
-
-	settingsJSON := fmt.Sprintf(`{"api_key":"test-key","model":"test-model"}`)
-	os.WriteFile(filepath.Join(tempDir, "settings.json"), []byte(settingsJSON), 0644)
-
-	job := Job{ArticleID: 301, Type: JobTypeEnrichFrontmatter}
-	pool.processEnrichFrontmatterWithURL(job, ts.URL)
-
-	// 1. Verify schema payload
-	if capturedResponseFormat == nil || capturedResponseFormat["type"] != "json_schema" {
-		t.Fatalf("expected response_format type json_schema, got: %v", capturedResponseFormat)
-	}
-	schemaDef, ok := capturedResponseFormat["json_schema"].(map[string]interface{})
-	if !ok || schemaDef["name"] != "okf_frontmatter" || schemaDef["strict"] != true {
-		t.Fatalf("expected strict json_schema okf_frontmatter, got: %v", schemaDef)
-	}
-
-	// 2. Verify file content and YAML parsing
-	updatedBytes, err := os.ReadFile(articlePath)
-	if err != nil {
-		t.Fatal(err)
-	}
-	updatedContent := string(updatedBytes)
-
-	if !strings.HasPrefix(updatedContent, "---\n") {
-		t.Fatalf("expected YAML frontmatter at start of file, got:\n%s", updatedContent)
-	}
-
-	parts := strings.SplitN(updatedContent, "---\n", 3)
-	if len(parts) < 3 {
-		t.Fatalf("expected frontmatter delimiter, got parts: %d", len(parts))
-	}
-
-	var parsed OKFMetadata
-	if err := yaml.Unmarshal([]byte(parts[1]), &parsed); err != nil {
-		t.Fatalf("failed to parse generated YAML frontmatter: %v", err)
-	}
-
-	if parsed.Type != "Playbook" {
-		t.Errorf("expected type 'Playbook', got %q", parsed.Type)
-	}
-	if parsed.Title != "Mastering Go Concurrency" {
-		t.Errorf("expected title 'Mastering Go Concurrency', got %q", parsed.Title)
-	}
-	if parsed.Description != "A comprehensive playbook for Go channels and goroutines." {
-		t.Errorf("expected description, got %q", parsed.Description)
-	}
-	if parsed.Resource != "https://golang.org/doc" {
-		t.Errorf("expected resource URL, got %q", parsed.Resource)
-	}
-	if len(parsed.Tags) != 3 || parsed.Tags[0] != "golang" {
-		t.Errorf("expected lowercase tags [golang, concurrency, channels], got: %v", parsed.Tags)
-	}
-	if parsed.Generated.By != "agent/readr-enricher" || parsed.Generated.At == "" {
-		t.Errorf("expected generated by agent/readr-enricher, got: %+v", parsed.Generated)
-	}
-
-	// 3. Verify body was preserved
-	if !strings.Contains(parts[2], "Go channels and goroutines make concurrent programming clean and safe.") {
-		t.Errorf("expected body to be preserved, got:\n%s", parts[2])
-	}
-}
