+++ backend/internal/agents/pipeline.go
@@ -19,12 +19,50 @@ import (
 	"gorm.io/gorm"
 )
 
+type llmLink struct {
+	ExistingArticleID int64  `json:"existing_article_id"`
+	ExactPhraseInText string `json:"exact_phrase_in_text"`
+}
+
+type OKFMetadata struct {
+	Type        string           `json:"type" yaml:"type"`
+	Title       string           `json:"title" yaml:"title"`
+	Description string           `json:"description" yaml:"description"`
+	Resource    string           `json:"resource,omitempty" yaml:"resource,omitempty"`
+	Tags        []string         `json:"tags" yaml:"tags"`
+	Generated   OKFGeneratedInfo `json:"-" yaml:"generated"`
+}
+
+type OKFGeneratedInfo struct {
+	By string `yaml:"by"`
+	At string `yaml:"at"`
+}
+
 type OKFFrontmatterResponse struct {
 	Type        string   `json:"type"`
 	Title       string   `json:"title"`
 	Description string   `json:"description"`
 	Resource    string   `json:"resource"`
 	Tags        []string `json:"tags"`
+}
+
+type jsonSchemaField struct {
+	Type                 string                     `json:"type"`
+	Properties           map[string]jsonSchemaField `json:"properties,omitempty"`
+	Items                *jsonSchemaField           `json:"items,omitempty"`
+	Required             []string                   `json:"required,omitempty"`
+	AdditionalProperties bool                       `json:"additionalProperties"`
+}
+
+type jsonSchemaDefinition struct {
+	Name   string          `json:"name"`
+	Strict bool            `json:"strict"`
+	Schema jsonSchemaField `json:"schema"`
+}
+
+type responseFormat struct {
+	Type       string                `json:"type"`
+	JSONSchema *jsonSchemaDefinition `json:"json_schema,omitempty"`
 }
 
 type UnifiedPipelineResponse struct {
@@ -39,6 +77,12 @@ type pipelineOpenRouterRequest struct {
 	ResponseFormat *responseFormat `json:"response_format,omitempty"`
 }
 
+type ArticleRecord struct {
+	ID    int64
+	Title string
+}
+
+var summaryBlockRegex = regexp.MustCompile(`(?m)^>\s*(?:💡\s*)?(?:\*\*)?(?:AI\s+)?Summary:(?:\*\*)?.*(?:\n>.*)*`)
 var reProtected = regexp.MustCompile(`(\[\[[\s\S]*?\]\]|\[[\s\S]*?\]\([\s\S]*?\)|` + "`[\\s\\S]*?`" + `)`)
 
 func injectLinksIntoBody(body string, links []llmLink, candidates []repository.ArticleRecord, sourceID int64, db *gorm.DB) (string, []string) {
