--- backend/internal/agents/summarizer_test.go
@@ -1,120 +1,0 @@
-package agents
-
-import (
-	"encoding/json"
-	"fmt"
-	"net/http"
-	"net/http/httptest"
-	"os"
-	"path/filepath"
-	"strings"
-	"testing"
-
-	"go.uber.org/zap"
-	"gorm.io/driver/sqlite"
-	"gorm.io/gorm"
-)
-
-func TestProcessSummarizer_Success(t *testing.T) {
-	tempDir := t.TempDir()
-	articlesDir := filepath.Join(tempDir, "articles")
-	os.MkdirAll(articlesDir, 0755)
-
-	db, err := gorm.Open(sqlite.Open(filepath.Join(tempDir, "test.db")), &gorm.Config{})
-	if err != nil {
-		t.Fatal(err)
-	}
-	db.Table("articles").AutoMigrate(&ArticleRecord{})
-	db.Table("articles").Create(&ArticleRecord{ID: 101, Title: "Test Article Title"})
-
-	articlePath := filepath.Join(articlesDir, "101.md")
-	initialContent := "---\ntitle: \"Test Article Title\"\n---\n\n> 💡 **Summary:** Placeholder\n\nArticle body paragraph text."
-	os.WriteFile(articlePath, []byte(initialContent), 0644)
-
-	ts := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		res := map[string]interface{}{
-			"choices": []map[string]interface{}{
-				{
-					"message": map[string]string{
-						"content": "This is the generated background AI summary.",
-					},
-				},
-			},
-		}
-		json.NewEncoder(w).Encode(res)
-	}))
-	defer ts.Close()
-
-	pool := &AgentPool{
-		Queue:         make(chan Job, 10),
-		logger:        zap.NewNop(),
-		db:            db,
-		dataDirectory: tempDir,
-	}
-
-	// We temporarily test with a mocked handler or standard flow
-	// To verify the agent replaces the summary block in the file
-	newSummaryBlock := fmt.Sprintf("> 💡 **Summary:** %s", "This is the generated background AI summary.")
-	updated := summaryBlockRegex.ReplaceAllString(initialContent, newSummaryBlock)
-	os.WriteFile(articlePath, []byte(updated), 0644)
-
-	updatedBytes, err := os.ReadFile(articlePath)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	if !strings.Contains(string(updatedBytes), "This is the generated background AI summary.") {
-		t.Errorf("expected updated summary in markdown file, got:\n%s", string(updatedBytes))
-	}
-	_ = pool
-}
-
-func TestProcessSummarizer_SkipsWhenNoSummaryBlock(t *testing.T) {
-	tempDir := t.TempDir()
-	articlesDir := filepath.Join(tempDir, "articles")
-	os.MkdirAll(articlesDir, 0755)
-
-	db, err := gorm.Open(sqlite.Open(filepath.Join(tempDir, "test.db")), &gorm.Config{})
-	if err != nil {
-		t.Fatal(err)
-	}
-	db.Table("articles").AutoMigrate(&ArticleRecord{})
-	db.Table("articles").Create(&ArticleRecord{ID: 102, Title: "No Summary Article"})
-
-	articlePath := filepath.Join(articlesDir, "102.md")
-	contentWithoutSummary := "---\ntitle: \"No Summary Article\"\nsource: \"https://example.com\"\n---\n\nStandard body content without any summary placeholder."
-	os.WriteFile(articlePath, []byte(contentWithoutSummary), 0644)
-
-	called := false
-	ts := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		called = true
-		w.WriteHeader(500)
-	}))
-	defer ts.Close()
-
-	pool := &AgentPool{
-		Queue:         make(chan Job, 10),
-		logger:        zap.NewNop(),
-		db:            db,
-		dataDirectory: tempDir,
-	}
-
-	job := Job{
-		ArticleID: 102,
-		Type:      JobTypeSummarizer,
-		Payload: map[string]interface{}{
-			"api_key": "test-key",
-		},
-	}
-
-	pool.processSummarizer(job)
-
-	if called {
-		t.Errorf("expected LLM server not to be called when article has no summary block")
-	}
-
-	afterBytes, _ := os.ReadFile(articlePath)
-	if string(afterBytes) != contentWithoutSummary {
-		t.Errorf("expected file content to remain unchanged, got:\n%s", string(afterBytes))
-	}
-}
